Release/2.5

+++ tools/testers/update-tester.sh
@@ -41,6 +41,24 @@ echo -
 
 CURRENT=2.5.0
 
+if [ ! -f build/sql/pgrouting--$CURRENT.sql ]; then
+   echo "File not found:  build/lib/pgrouting--$CURRENT.sql "
+   exit 1
+fi
+
+dropdb --if-exists ___test_update
+
+#make sure the scripts are the latest
+cp build/sql/pgrouting-* tools/sql-update-scripts
+#TODO automatically check  the diff is empty
+#git diff
+
+cd build
+sudo make install
+cd ..
+
+
+
 function update_test {
 
 echo ------------------------------------
@@ -54,33 +72,52 @@ if [ "$INSTALLED" == "/usr/share/postgresql/9.3/extension/pgrouting--$1.sql" ]
 then
     echo "/usr/share/postgresql/9.3/extension/pgrouting--$1.sql found"
 else
-    echo "/usr/share/postgresql/9.3/extension/pgrouting--$1.sql Not found"
+    echo "FATAL: /usr/share/postgresql/9.3/extension/pgrouting--$1.sql Not found"
     exit 1
 fi
 
+
 createdb  ___test_update
 psql  ___test_update  <<EOF
-create extension postgis;
-create extension pgrouting with version '$1';
-select pgr_version();
-alter extension pgrouting update to '$2';
-select pgr_version();
+CREATE extension postgis;
+CREATE extension pgrouting with version '$1';
 EOF
 
-dropdb   ___test_update
-} 
+OLD_VERSION=$(psql ___test_update -t -c 'SELECT version FROM pgr_version()')
+
+
+if [ "b$OLD_VERSION" != "b $1" ]
+then
+    echo "ERROR: Version $1 not found on the system"
+    dropdb ___test_update
+    exit 1
+fi
+
+
+
+psql ___test_update -c "ALTER extension pgrouting update to '$2'"
+
+
+NEW_VERSION=$(psql ___test_update -t -c 'SELECT version FROM pgr_version()')
+
+echo "$OLD_VERSION ->> $NEW_VERSION"
+
+if [ "b$NEW_VERSION" != "b $2" ]
+then
+    echo "FAIL: Could not update from version $1 to version $2"
+    dropdb ___test_update
+    exit 1
+fi
+
+dropdb ___test_update
+
+} # end of function
 
 #------------------------------------
 ### updates from 2.4
 #------------------------------------
 
 update_test 2.4.1 $CURRENT
-update_test 2.4.0 $CURRENT
-
-#------------------------------------
-### updates from 2.3.0
-#------------------------------------
-
 update_test 2.4.0 $CURRENT
 
 #------------------------------------
@@ -95,7 +132,7 @@ update_test 2.3.0 $CURRENT
 ### updates from 2.2.x
 #------------------------------------
 
-#update_test 2.2.4 $CURRENT
+update_test 2.2.4 $CURRENT
 update_test 2.2.3 $CURRENT
 update_test 2.2.2 $CURRENT
 update_test 2.2.1 $CURRENT
